test: SessionManager 동시성 테스트 패키지 이동
